Fixed file size format

+++ src/Nethermind.Packager.Core/Services/Dto/PackageDto.cs
@@ -8,7 +8,7 @@ namespace Nethermind.Packager.Core.Services.Dto
         public string Commit { get; set; }
         public string Kind { get; set; }
         public string Arch { get; set; }
-        public long Size { get; set; }
+        public double Size { get; set; }
         public DateTime PublishedAt { get; set; }
         public string Signature { get; set; }
         public string Checksum { get; set; }
